chore: 新增发 nightly 版本的 workflow

+++ tools/OTAPacker/build.sh
@@ -2,10 +2,7 @@
 
 source_repo=$1
 releases_txt="$2"
-
-ghrepo() {
-    gh $@ --repo $source_repo
-}
+source_repo_fallback=$3
 
 working_dir="$(pwd)"
 echo "working_dir: $working_dir"
@@ -15,18 +12,19 @@ script_dir="$(dirname $(realpath "$0"))"
 latest_tag=$(head -n 1 "$releases_txt")
 
 while read tag; do
-    (
+    if [ ! -d "$tag"/content ]; then
         mkdir -pv "$tag"
         cd "$tag"
         echo "Downloading $tag"
-        ghrepo release download "$tag" --pattern "MAA-$tag-win-x64.zip" --pattern "MaaBundle-$tag.zip" --clobber
+        gh release download "$tag" --repo $source_repo --pattern "MAA-$tag-win-x64.zip" --pattern "MaaBundle-$tag.zip" --clobber \
+            || gh release download "$tag" --repo $source_repo_fallback --pattern "MAA-$tag-win-x64.zip" --pattern "MaaBundle-$tag.zip" --clobber
         mkdir -pv 'content'
-        echo "Unzip" *.zip
-        unzip -qq -O gbk -o "*.zip" -d 'content'
+        unzip -q -O gbk -o "*.zip" -d 'content'
         rm -fv *.zip
-    )
-
-    cd $working_dir
+        cd $working_dir
+    else
+        echo "$tag"/content/ already exists
+    fi
 
     if [[ "$tag" == "$latest_tag" ]]; then
         cd "$latest_tag"/content/
